New caching implementation.

+++ src/competition/cig/desimonenotarangelo/ScoreEvaluatorAgent/NeuralNetwork/NeuralNetworkInput.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface NeuralNetworkInput
 {
-  public List<Byte> getInputList();
+  public List<Double> getInputList();
   public int size();
 }
